REFACT - Fixed dependencies for independent modules - added dependencies between modules where required; - moved HTML classes to XML module to avoid circular dependency; - created explicit module directory for UTF-8 module; and - moved unit tests and test resources to specific modules.
