+++ jhove-modules/html-hul/src/main/java/edu/harvard/hul/ois/jhove/module/HtmlModule.java
@@ -23,6 +23,7 @@ import java.io.*;
 import java.util.*;
 import edu.harvard.hul.ois.jhove.*;
 import edu.harvard.hul.ois.jhove.module.html.*;
+import edu.harvard.hul.ois.jhove.module.xml.HtmlMetadata;
 
 /**
  * Module for identification and validation of HTML files.
