+++ jhove-modules/tiff-hul/src/test/java/edu/harvard/hul/ois/jhove/module/tiff/TiffProfileExifIFDTest.java
